Refactor: mvc 구조로 프로젝트 설계 변경
